Initial PoC codedrop
See README for the many things still wrong with this. Process
separation on Android complicates things a lot. The original code was
written under the assumption that we can have the UI application
accept HTTP requests in the background - after it became clear that
needs to be in an Android foreground service things became
significantly more complicated, and some initial design decisions
slightly problematic. There's still some work to be done to fix this.