Attempt to work around StackView's component caching
StackView, when told to use a string-based source, interprets it as a
component path, and creates a QQmlComponent instance which it takes
ownership of. This is good for the typical case where components are
static, and unchanging, but we are a special delicate flower, and want
dynamicism, so for us it is a problem.

The workaround, then: we create a QQmlComponent ourselves, instantiate
an item from it, and push() that QQuickItem, destroying the component
when we are done with it.