Add separate log reader
emacsbridge -log can now be used to read the internal ringbuffer with
log data from the server.

The client now prefers logging to the server through the remote
object, so the majority of the client side log messages should also be
available in the server log.

This doesn't work properly on Android yet, though stubs are present
for that as well.