Fix directory listings and add /local address
This now allows proper recursing of directories, so the "list all on
the first page" workaround is no longer required.

The /local address can be used to serve local files if configured in
the server settings - this currently is mostly useful for testing
script/lisp download without having to rebuild the binary.

Nice to have would be correct mime types for the templated files as
well as fancy listing with icons, modification dates and sizes - as
that'd mostly make the /local usage more comfortable it's not really
urgent, though.